Move GpgCheck to Verify::Gpg, keep disabled
